// src/app/lib/generated/models/Ethnicity.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AI Babe Public API
 * Lorem ipsum
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const Ethnicity = {
    Asian: 'asian',
    Black: 'black',
    Caucasian: 'caucasian',
    Arab: 'arab',
    Latina: 'latina',
    Indian: 'indian'
} as const;
export type Ethnicity = typeof Ethnicity[keyof typeof Ethnicity];


export function instanceOfEthnicity(value: any): boolean {
    for (const key in Ethnicity) {
        if (Object.prototype.hasOwnProperty.call(Ethnicity, key)) {
            if (Ethnicity[key as keyof typeof Ethnicity] === value) {
                return true;
            }
        }
    }
    return false;
}

export function EthnicityFromJSON(json: any): Ethnicity {
    return EthnicityFromJSONTyped(json, false);
}

export function EthnicityFromJSONTyped(json: any, ignoreDiscriminator: boolean): Ethnicity {
    return json as Ethnicity;
}

export function EthnicityToJSON(value?: Ethnicity | null): any {
    return value as any;
}

export function EthnicityToJSONTyped(value: any, ignoreDiscriminator: boolean): Ethnicity {
    return value as Ethnicity;
}